# Conflicts:
#	src/lib/operators/aggregate.cpp

+++ src/lib/storage/table.cpp
@@ -18,7 +18,7 @@
 
 namespace opossum {
 
-Table::Table(const size_t chunk_size) : _chunk_size(chunk_size), _append_mutex(std::make_unique<std::mutex>()) {
+Table::Table(const uint32_t chunk_size) : _chunk_size(chunk_size), _append_mutex(std::make_unique<std::mutex>()) {
   _chunks.push_back(Chunk{true});
 }
 
@@ -60,8 +60,8 @@ void Table::create_new_chunk() {
 
 uint16_t Table::col_count() const { return _column_types.size(); }
 
-uint32_t Table::row_count() const {
-  size_t ret = 0;
+uint64_t Table::row_count() const {
+  uint64_t ret = 0;
   for (auto &&chunk : _chunks) {
     ret += chunk.size();
   }
